Fix GetTokenAndMetaDataFromFunction riid direction and GetNativeCodeStartAddresses pointer size
Fix two binding bugs in the profiler COM interface declarations:

1. GetTokenAndMetaDataFromFunction: Change riid parameter from 'out Guid' to
   'in Guid'. The REFIID is an input specifying which metadata interface the
   caller wants (e.g. IID_IMetaDataImport), not an output. The incorrect 'out'
   direction caused the CLR to receive garbage as the IID, returning
   E_NOINTERFACE. The wrapper now accepts riid as a parameter and TokenAndMetaData
   no longer includes the redundant Riid field.

2. GetNativeCodeStartAddresses: Change code start address type from uint/uint*
   to nint/nint*. Native code addresses are UINT_PTR (pointer-sized), so using
   32-bit uint truncated 64-bit addresses on x64 systems.

Add FunctionInfoTests that validate both fixes:
- GetTokenAndMetaDataFromFunction: cross-validates the returned token against
  GetFunctionInfo, then uses the returned IMetaDataImport pointer to call
  GetMethodProps, proving the interface pointer is valid.
- GetNativeCodeStartAddresses: passes each returned address to GetCodeInfo4,
  proving the addresses are not truncated.

+++ src/Silhouette/Interfaces/ICorProfilerInfo9.cs
@@ -6,7 +6,7 @@ internal unsafe interface ICorProfilerInfo9 : ICorProfilerInfo8
     public new static readonly Guid Guid = new("008170DB-F8CC-4796-9A51-DC8AA0B47012");
 
     //Given functionId + rejitId, enumerate the native code start address of all jitted versions of this code that currently exist
-    HResult GetNativeCodeStartAddresses(FunctionId functionID, ReJITId reJitId, uint cCodeStartAddresses, out uint pcCodeStartAddresses, uint* codeStartAddresses);
+    HResult GetNativeCodeStartAddresses(FunctionId functionID, ReJITId reJitId, uint cCodeStartAddresses, out uint pcCodeStartAddresses, nint* codeStartAddresses);
 
     //Given the native code start address, return the native->IL mapping information for this jitted version of the code
     HResult GetILToNativeMapping3(nint nativeCodeStartAddress, uint cMap, out uint pcMap, COR_DEBUG_IL_TO_NATIVE_MAP* map);
